prismx/serializers.py
 
from rest_framework import serializers

from prismx.models import *

from django.contrib.auth import get_user_model

User = get_user_model()

class AccountSerializer(serializers.ModelSerializer):

    class Meta:
        model = Account
        fields = '__all__'