requirements.txt
CherryPy==18.6.0
ConfigUpdater==1.1.3
Django==3.0.10
Mako==1.0.7
MarkupSafe==1.1.0
Pillow==5.1.0
PyJWT==1.7.1
PyNaCl==1.1.2
PyYAML
ScoutSuite
SecretStorage==2.3.1
adal==1.2.4
aliyun_python_sdk_actiontrail==2.0.1
aliyun_python_sdk_core==2.13.26
aliyun_python_sdk_ecs==4.19.12
aliyun_python_sdk_kms==2.11.0
aliyun_python_sdk_ram==3.2.0
aliyun_python_sdk_rds==2.5.1
aliyun_python_sdk_sts==3.0.2
aliyun_python_sdk_vpc==3.0.11
amqp==2.6.0
asgiref==3.2.10
asn1crypto==0.24.0
asyncio-throttle==1.0.1
atlassian_python_api==1.16.1
beautifulsoup4==4.8.2
billiard==3.6.3.0
boto3==1.14.60
botocore==1.17.60
celery
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cheroot==8.3.0
cherrypy_cors==1.6
coloredlogs==14.0
cryptography==2.9.2
dj-database-url==0.5.0
django-celery-beat
django-celery-results
django-widget-tweaks
djangorestframework==3.11.1
docutils==0.15.2
future==0.18.2
google_api_python_client==1.12.2
gunicorn==20.0.4
httpagentparser==1.9.0
httplib2==0.11.3
httplib2shim==0.0.3
humanfriendly==8.2
idna==2.6
importlib-metadata==1.7.0
importlib-resources==3.0.0
jaraco.classes==3.1.0
jaraco.collections==3.0.0
jaraco.functools==3.0.1
jaraco.text==3.2.0
jmespath==0.10.0
keyring==10.6.0
keyrings.alt==3.0
kombu==4.6.11
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
linecache2==1.0.0
more-itertools==8.4.0
msrestazure==0.6.4
netaddr==0.8.0
oauth==1.0.1
olefile==0.45.1
oss2==2.13.0
pbr==5.4.4
pexpect==4.2.1
policyuniverse==1.3.2.3
portend==2.6
protobuf==3.13.0
psutil==5.6.7
psycopg2==2.8.5
ptyprocess==0.6.0
pyRFC3339==1.0
pycparser==2.20
pymacaroons==0.13.0
python-debian==0.1.32
python-xlib==0.26
python_dateutil==2.8.1
pytz==2019.3
pyxdg==0.25
reportlab==3.4.0
requests-unixsocket==0.1.5
requests==2.18.4
s3transfer==0.3.3
selenium==3.141.0
setproctitle==1.1.10
simplejson==3.13.2
six==1.13.0
soupsieve==2.0.1
sqlitedict==1.7.0
sqlparse==0.3.0
tempora==3.0.0
testresources==2.0.1
urllib3==1.22
vine==1.3.0
wadllib==1.3.4
whitenoise==5.1.0
zc.lockfile==2.0
zipp==3.1.0
